+++ components/Main/Splash/ScrollToCategory.tsx
