+++ components/Main/Splash/index.stories.tsx
