+++ components/common/Banner/Banner.tsx
