+++ constants/dummyData/banner.ts
@@ -4,24 +4,24 @@ export const BANNER_DUMMY_DATA: BannerInformation[] = [
   {
     background: '#232328',
     img: {
-      default: '/asset/banner/monitor_banner_01.png',
-      phone: '/asset/banner/phone_banner_01.png'
+      default: '/images/banner/monitor_banner_01.png',
+      phone: '/images/banner/phone_banner_01.png'
     },
     description: '냉장고에 뭐가 있더라'
   },
   {
     background: '#232328',
     img: {
-      default: '/asset/banner/monitor_banner_01.png',
-      phone: '/asset/banner/phone_banner_01.png'
+      default: '/images/banner/monitor_banner_01.png',
+      phone: '/images/banner/phone_banner_01.png'
     },
     description: '냉장고에 뭐가 있더라'
   },
   {
     background: '#232328',
     img: {
-      default: '/asset/banner/monitor_banner_01.png',
-      phone: '/asset/banner/phone_banner_01.png'
+      default: '/images/banner/monitor_banner_01.png',
+      phone: '/images/banner/phone_banner_01.png'
     },
     description: '냉장고에 뭐가 있더라'
   }
