+++ pages/_app.tsx
@@ -2,7 +2,7 @@ import theme from '@/styles/theme';
 import { Global, ThemeProvider } from '@emotion/react';
 import type { AppProps } from 'next/app';
 import { global } from '@/styles/Global';
-import ReactQueryProvider from '@/components/react-query/ReactQueryProvider';
+import ReactQueryProvider from '@/provider/ReactQueryProvider';
 import Layout from '@/components/layout/Layout';
 // Swiper
 import 'swiper/css';
